pile of minor fixes
don't set a variable then return the variable.

del variables when we're done w/ them so that gc knows to kill it.

called self.verify_state() and not some global lm.verify_state()

import secrets at the top of the file not in a function.

make a log slightly clearer